Correct usage of sqe->user_data
This is expicitly a 64bit value. It would be nice if this were
documented, especially in the examples that use it by packing a struct
(but not striclty packing it) to be a 32 bit value and a 16 bit value.

This means that we need to pass pointers to the info structs, not try
and stash them because we are 128 bits with the to_free pointer.

Signed-off-by: Eric B Munson <munsoner@bu.edu>